RIA-7244 Merge pact tests
