--- src/contractTest/resources/commonrefdata/CommonRefData.json
@@ -1,30 +1,0 @@
-{
-  "list_of_values": [
-    {
-      "category_key": "InterpreterLanguage",
-      "key" : "a1fr",
-      "value_en": "Afrikaans",
-      "value_cy": " ",
-      "hint_text_en": "",
-      "hint_text_cy": "",
-      "lov_order": "null",
-      "parent_category": "",
-      "parent_key": "",
-      "active_flag": "Y",
-      "child_nodes": null
-    },
-    {
-      "category_key": "InterpreterLanguage",
-      "key" : "arq",
-      "value_en": "Algerian",
-      "value_cy": " ",
-      "hint_text_en": "",
-      "hint_text_cy": "",
-      "lov_order": "null",
-      "parent_category": "",
-      "parent_key": "",
-      "active_flag": "Y",
-      "child_nodes": null
-    }
-  ]
-}
