+++ src/main/java/uk/gov/hmcts/reform/iacaseapi/infrastructure/clients/refdata/CommonDataRefApi.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.iacaseapi.infrastructure.clients;
+package uk.gov.hmcts.reform.iacaseapi.infrastructure.clients.refdata;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.cloud.openfeign.FeignClientProperties;
